folder structure changed for navigations
