+++ App.js
@@ -1,10 +1,5 @@
 import { StatusBar } from 'expo-status-bar';
-import { StyleSheet } from 'react-native';
 import { NavigationContainer } from '@react-navigation/native';
-import { createNativeStackNavigator } from '@react-navigation/native-stack';
-import Register from './src/screens/Register';
-import Login from './src/screens/Login';
-import Dashboard from './src/screens/Dashboard';
 import { useFonts } from 'expo-font';
 import AppLoading from 'expo-app-loading';
 import {
@@ -27,10 +22,7 @@ import {
   Poppins_900Black,
   Poppins_900Black_Italic,
 } from '@expo-google-fonts/poppins';
-import Splash from './src/screens/Splash';
-import { useState } from 'react';
-
-const Stack = createNativeStackNavigator();
+import StackNavigation from './src/navigations/StackNavigation';
 
 export default function App() {
   let [fontsLoaded, error] = useFonts({
@@ -53,42 +45,14 @@ export default function App() {
     Poppins_900Black,
     Poppins_900Black_Italic,
   });
-  let [isReady, setIsReady] = useState(false);
   if (!fontsLoaded) {
     return <AppLoading />;
   }
 
   return (
     <NavigationContainer>
-      <Stack.Navigator initialRouteName='Splash'>
-        <Stack.Screen
-          options={{ headerShown: false }}
-          name='Splash'
-          component={Splash}
-        />
-        <Stack.Screen
-          options={{ headerShown: false }}
-          name='Sign Up'
-          component={Register}
-        />
-        <Stack.Screen
-          options={{ headerShown: false }}
-          name='Sign In'
-          component={Login}
-        />
-        <Stack.Screen
-          options={{ headerShown: false }}
-          name='Dashboard'
-          component={Dashboard}
-        />
-      </Stack.Navigator>
+      <StackNavigation />
       <StatusBar style='auto' />
     </NavigationContainer>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-  },
-});
